+++ include/ssGUI/Backend/SFML/BackendDrawingSFML.hpp
@@ -32,38 +32,38 @@ namespace ssGUI::Backend
             int BackendIndex; 
 
             //Non index variants exist only for legacy purposes
-            bool DrawShape( const std::vector<glm::ivec2>& vertices, 
-                            const std::vector<glm::ivec2>& texCoords,
+            bool DrawShape( const std::vector<glm::vec2>& vertices, 
+                            const std::vector<glm::vec2>& texCoords,
                             const std::vector<glm::u8vec4>& colours,
                             const ssGUI::Backend::BackendFontInterface& font,
                             int characterSize) override;
 
-            bool DrawShape( const std::vector<glm::ivec2>& vertices, 
-                            const std::vector<glm::ivec2>& texCoords,
+            bool DrawShape( const std::vector<glm::vec2>& vertices, 
+                            const std::vector<glm::vec2>& texCoords,
                             const std::vector<glm::u8vec4>& colours,
                             const ssGUI::Backend::BackendImageInterface& image) override;
 
-            bool DrawShape( const std::vector<glm::ivec2>& vertices, 
+            bool DrawShape( const std::vector<glm::vec2>& vertices, 
                             const std::vector<glm::u8vec4>& colours) override;
 
             //NOTE: End index is exclusive
-            bool DrawShape( const std::vector<glm::ivec2>& vertices, 
-                                    const std::vector<glm::ivec2>& texCoords,
-                                    const std::vector<glm::u8vec4>& colours,
-                                    int startIndex, int endIndex,
-                                    const ssGUI::Backend::BackendFontInterface& font,
-                                    int CharacterSize) override;
+            bool DrawShape( const std::vector<glm::vec2>& vertices, 
+                            const std::vector<glm::vec2>& texCoords,
+                            const std::vector<glm::u8vec4>& colours,
+                            int startIndex, int endIndex,
+                            const ssGUI::Backend::BackendFontInterface& font,
+                            int CharacterSize) override;
 
-            bool DrawShape( const std::vector<glm::ivec2>& vertices, 
-                                    const std::vector<glm::ivec2>& texCoords,
-                                    const std::vector<glm::u8vec4>& colours,
-                                    int startIndex, int endIndex,
-                                    const ssGUI::Backend::BackendImageInterface& image) override;
+            bool DrawShape( const std::vector<glm::vec2>& vertices, 
+                            const std::vector<glm::vec2>& texCoords,
+                            const std::vector<glm::u8vec4>& colours,
+                            int startIndex, int endIndex,
+                            const ssGUI::Backend::BackendImageInterface& image) override;
 
 
-            bool DrawShape( const std::vector<glm::ivec2>& vertices, 
-                                    const std::vector<glm::u8vec4>& colours,
-                                    int startIndex, int endIndex) override;
+            bool DrawShape( const std::vector<glm::vec2>& vertices, 
+                            const std::vector<glm::u8vec4>& colours,
+                            int startIndex, int endIndex) override;
 
         protected:
             
@@ -74,8 +74,8 @@ namespace ssGUI::Backend
 
             //function: DrawEntities
             //See <BackendDrawingInterface::DrawEntities>
-            bool DrawEntities(  const std::vector<glm::ivec2>& vertices, 
-                                const std::vector<glm::ivec2>& texCoords,
+            bool DrawEntities(  const std::vector<glm::vec2>& vertices, 
+                                const std::vector<glm::vec2>& texCoords,
                                 const std::vector<glm::u8vec4>& colours,
                                 const std::vector<int>& counts,
                                 const std::vector<ssGUI::DrawingProperty>& properties) override;
